Add dataset nodes to dag graph
